// web/src/types/api.ts
export type EsimProfileItem = {
  iccid: string
  name: string
  service_provider_name: string
  state: number
  state_text: string
  class_text?: string
}

export type EsimEUICCProfiles = {
  eid: string
  aid_hex: string
  profiles: EsimProfileItem[]
}